requirements.txt
Flask==2.1.
Werkzeug==2.2.2
mysql-connector-python
Flask-Cors>=3.0.0
mysql-connector-python==8.0.28
requests==2.26.0

